+++ src/components/SQFormDialog/SQFormDialog.tsx
@@ -29,7 +29,9 @@ interface SQFormDialogProps<Values> {
     values: Values,
     formikHelpers: FormikHelpers<Values>
   ) => void | Promise<unknown>;
-  /** Whether to show save/submit button */
+  /** Determine if the secondary action button should be displayed (default: true) */
+  showSecondaryButton?: boolean;
+  /** Whether to show save/submit button (default: true) */
   shouldDisplaySaveButton?: boolean;
   /** The primary button text (Button located on right side of Dialog) */
   saveButtonText?: string;
@@ -62,6 +64,7 @@ function SQFormDialog<Values>({
   maxWidth = 'sm',
   onClose,
   onSave,
+  showSecondaryButton = true,
   shouldDisplaySaveButton = true,
   saveButtonText = 'Save',
   title,
@@ -104,6 +107,7 @@ function SQFormDialog<Values>({
         shouldRequireFieldUpdates={shouldRequireFieldUpdates}
         title={title}
         muiGridProps={muiGridProps}
+        showSecondaryButton={showSecondaryButton}
       />
     </Formik>
   );