+++ src/components/SQFormGuidedWorkflow/SQFormGuidedWorkflow.js
@@ -52,6 +52,7 @@ function SQFormGuidedWorkflow({
   initialCompletedTasks = 0,
   isStrictMode = false,
   onError,
+  containerStyles = {},
 }) {
   const classes = useStyles();
 
@@ -171,7 +172,7 @@ function SQFormGuidedWorkflow({
   });
 
   return (
-    <Section style={{padding: '20px'}}>
+    <Section style={{padding: '20px', ...containerStyles}}>
       <SectionHeader title={mainTitle} informativeHeading={mainSubtitle} />
       <SectionBody>
         <Accordion accordionPanels={transformedTaskModules} />