+++ stories/__tests__/SQFormDialog.stories.test.js
@@ -18,7 +18,7 @@ afterEach(() => {
 });
 
 const mockData = {
-  hello: 'howdy'
+  hello: 'howdy',
 };
 
 describe('Tests for Default', () => {
@@ -87,13 +87,49 @@ describe('Tests for Default', () => {
     // https://testing-library.com/docs/dom-testing-library/api-events/ --> find 'keyboard events'
     fireEvent.keyDown(screen.getByRole('presentation'), {
       key: 'Escape',
-      code: 'Escape'
+      code: 'Escape',
     });
 
     await waitFor(() => {
       expect(handleClose).toHaveBeenCalledTimes(1);
     });
   });
+});
+
+it('should not call onClose on `escape` keydown because cancel is not available', async () => {
+  render(
+    <Default
+      isOpen={true}
+      onSave={handleSave}
+      onClose={handleClose}
+      showSecondaryButton={false}
+    />
+  );
+
+  // fireEvent, not userEvent
+  // to confirm the 'key' and 'code' values-- > https://keycode.info/
+  // https://testing-library.com/docs/dom-testing-library/api-events/ --> find 'keyboard events'
+  fireEvent.keyDown(screen.getByRole('presentation'), {
+    key: 'Escape',
+    code: 'Escape',
+  });
+
+  await waitFor(() => {
+    expect(handleClose).not.toHaveBeenCalled();
+  });
+});
+
+it('should not find the cancel secondary button', async () => {
+  render(
+    <Default
+      isOpen={true}
+      onSave={handleSave}
+      onClose={handleClose}
+      showSecondaryButton={false}
+    />
+  );
+
+  expect(screen.queryByRole('button', {name: /cancel/i})).toBeNull();
 });
 
 describe('Tests for WithValidation', () => {